[fli-iam#1867] Delete useless interface
