+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/vip/monitoring/schedule/ExecutionStatusMonitorRunner.java
@@ -23,7 +23,7 @@ public class ExecutionStatusMonitorRunner implements ApplicationRunner {
     private ExecutionMonitoringService execMonitoringSrv;
 
     @Autowired
-    private ExecutionStatusMonitor execMonitor;
+    private ExecutionStatusMonitorService execMonitor;
 
 
     /**
